requirements.txt
azure-mgmt-datafactory
azure-identity
azure-mgmt-resource
azure-mgmt-storage
python-dotenv
azure-storage-blob